refactor: move usecase/model package to domain/model
