+++ app/src/main/java/kr/ksw/visitkorea/data/local/entity/FavoriteEntity.kt
@@ -1,0 +1,3 @@
+package kr.ksw.visitkorea.data.local.entity
+
+data class FavoriteEntity()
