+++ app/src/main/java/kr/ksw/visitkorea/domain/model/CommonDetail.kt
@@ -1,4 +1,4 @@
-package kr.ksw.visitkorea.domain.usecase.model
+package kr.ksw.visitkorea.domain.model
 
 data class CommonDetail(
     val parking: String? = null,
