+++ app/src/main/java/kr/ksw/visitkorea/domain/model/HotelRoomDetail.kt
@@ -1,4 +1,4 @@
-package kr.ksw.visitkorea.domain.usecase.model
+package kr.ksw.visitkorea.domain.model
 
 data class HotelRoomDetail(
     val roomTitle: String = "",
