+++ app/src/main/java/kr/ksw/visitkorea/domain/model/TouristSpot.kt
@@ -1,4 +1,4 @@
-package kr.ksw.visitkorea.domain.usecase.model
+package kr.ksw.visitkorea.domain.model
 
 /**
  * For Home Screen Card Model
