+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/detail/GetDetailIntroUseCase.kt
@@ -1,6 +1,6 @@
 package kr.ksw.visitkorea.domain.usecase.detail
 
-import kr.ksw.visitkorea.domain.usecase.model.CommonDetail
+import kr.ksw.visitkorea.domain.model.CommonDetail
 
 interface GetDetailIntroUseCase {
     suspend operator fun invoke(
