+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/detail/GetDetailIntroUseCaseImpl.kt
@@ -2,7 +2,7 @@ package kr.ksw.visitkorea.domain.usecase.detail
 
 import kr.ksw.visitkorea.data.repository.DetailRepository
 import kr.ksw.visitkorea.domain.usecase.mapper.toCommonDetail
-import kr.ksw.visitkorea.domain.usecase.model.CommonDetail
+import kr.ksw.visitkorea.domain.model.CommonDetail
 import javax.inject.Inject
 
 class GetDetailIntroUseCaseImpl @Inject constructor(
