+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/detail/GetHotelDetailUseCase.kt
@@ -1,6 +1,6 @@
 package kr.ksw.visitkorea.domain.usecase.detail
 
-import kr.ksw.visitkorea.domain.usecase.model.HotelDetail
+import kr.ksw.visitkorea.domain.model.HotelDetail
 
 interface GetHotelDetailUseCase {
     suspend operator fun invoke(
