+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/detail/GetHotelDetailUseCaseImpl.kt
@@ -3,7 +3,7 @@ package kr.ksw.visitkorea.domain.usecase.detail
 import kr.ksw.visitkorea.data.repository.DetailRepository
 import kr.ksw.visitkorea.domain.common.TYPE_HOTEL
 import kr.ksw.visitkorea.domain.usecase.mapper.toHotelDetail
-import kr.ksw.visitkorea.domain.usecase.model.HotelDetail
+import kr.ksw.visitkorea.domain.model.HotelDetail
 import javax.inject.Inject
 
 class GetHotelDetailUseCaseImpl @Inject constructor(
