+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/detail/GetHotelRoomDetailUseCase.kt
@@ -1,6 +1,6 @@
 package kr.ksw.visitkorea.domain.usecase.detail
 
-import kr.ksw.visitkorea.domain.usecase.model.HotelRoomDetail
+import kr.ksw.visitkorea.domain.model.HotelRoomDetail
 
 interface GetHotelRoomDetailUseCase {
     suspend operator fun invoke(
