+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/detail/GetHotelRoomDetailUseCaseImpl.kt
@@ -2,7 +2,7 @@ package kr.ksw.visitkorea.domain.usecase.detail
 
 import kr.ksw.visitkorea.data.repository.DetailRepository
 import kr.ksw.visitkorea.domain.usecase.mapper.toHotelRoomDetail
-import kr.ksw.visitkorea.domain.usecase.model.HotelRoomDetail
+import kr.ksw.visitkorea.domain.model.HotelRoomDetail
 import javax.inject.Inject
 
 class GetHotelRoomDetailUseCaseImpl @Inject constructor(
