+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/home/GetTouristSpotForHomeUseCase.kt
@@ -1,6 +1,6 @@
 package kr.ksw.visitkorea.domain.usecase.home
 
-import kr.ksw.visitkorea.domain.usecase.model.TouristSpot
+import kr.ksw.visitkorea.domain.model.TouristSpot
 
 interface GetTouristSpotForHomeUseCase {
     suspend operator fun invoke(
