+++ app/src/main/java/kr/ksw/visitkorea/domain/usecase/mapper/ToPresentationModelMapper.kt
@@ -8,14 +8,14 @@ import kr.ksw.visitkorea.domain.common.TYPE_CULTURE
 import kr.ksw.visitkorea.domain.common.TYPE_LEiSURE
 import kr.ksw.visitkorea.domain.common.TYPE_RESTAURANT
 import kr.ksw.visitkorea.domain.common.TYPE_TOURIST_SPOT
-import kr.ksw.visitkorea.domain.usecase.model.CommonCardModel
-import kr.ksw.visitkorea.domain.usecase.model.CommonDetail
-import kr.ksw.visitkorea.domain.usecase.model.Festival
-import kr.ksw.visitkorea.domain.usecase.model.HotelDetail
-import kr.ksw.visitkorea.domain.usecase.model.HotelRoomDetail
-import kr.ksw.visitkorea.domain.usecase.model.MoreCardModel
-import kr.ksw.visitkorea.domain.usecase.model.Restaurant
-import kr.ksw.visitkorea.domain.usecase.model.TouristSpot
+import kr.ksw.visitkorea.domain.model.CommonCardModel
+import kr.ksw.visitkorea.domain.model.CommonDetail
+import kr.ksw.visitkorea.domain.model.Festival
+import kr.ksw.visitkorea.domain.model.HotelDetail
+import kr.ksw.visitkorea.domain.model.HotelRoomDetail
+import kr.ksw.visitkorea.domain.model.MoreCardModel
+import kr.ksw.visitkorea.domain.model.Restaurant
+import kr.ksw.visitkorea.domain.model.TouristSpot
 import kr.ksw.visitkorea.domain.usecase.util.toDateString
 import kr.ksw.visitkorea.domain.usecase.util.toDistForUi
 import kr.ksw.visitkorea.domain.usecase.util.toImageUrl
