+++ app/src/main/java/kr/ksw/visitkorea/presentation/detail/component/DetailHotelCard.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import kr.ksw.visitkorea.R
-import kr.ksw.visitkorea.domain.usecase.model.HotelRoomDetail
+import kr.ksw.visitkorea.domain.model.HotelRoomDetail
 import kr.ksw.visitkorea.presentation.ui.theme.VisitKoreaTheme
 
 @OptIn(ExperimentalLayoutApi::class)
