+++ app/src/main/java/kr/ksw/visitkorea/presentation/festival/viewmodel/FestivalState.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Immutable
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-import kr.ksw.visitkorea.domain.usecase.model.Festival
+import kr.ksw.visitkorea.domain.model.Festival
 
 @Immutable
 data class FestivalState(
