+++ app/src/main/java/kr/ksw/visitkorea/presentation/hotel/screen/HotelScreen.kt
@@ -30,7 +30,7 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import kotlinx.coroutines.flow.collectLatest
 import kr.ksw.visitkorea.domain.common.TYPE_HOTEL
-import kr.ksw.visitkorea.domain.usecase.model.CommonCardModel
+import kr.ksw.visitkorea.domain.model.CommonCardModel
 import kr.ksw.visitkorea.presentation.common.DetailParcel
 import kr.ksw.visitkorea.presentation.detail.DetailActivity
 import kr.ksw.visitkorea.presentation.home.component.CultureCard
