+++ app/src/main/java/kr/ksw/visitkorea/presentation/hotel/viewmodel/HotelState.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Immutable
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-import kr.ksw.visitkorea.domain.usecase.model.CommonCardModel
+import kr.ksw.visitkorea.domain.model.CommonCardModel
 
 @Immutable
 data class HotelState(
