+++ app/src/main/java/kr/ksw/visitkorea/presentation/more/viewmodel/MoreState.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Immutable
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-import kr.ksw.visitkorea.domain.usecase.model.MoreCardModel
+import kr.ksw.visitkorea.domain.model.MoreCardModel
 
 @Immutable
 data class MoreState(
